refactor: move book_invitation_live
